# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TodoPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List withdrawal requests
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: approvalStatus
          in: query
          schema:
            type: string
        - name: withdrawalStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalList"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      summary: Approve withdrawal request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals]
      summary: Reject withdrawal request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    Withdrawal:
      type: object
      required: [id, trackingNumber, amount, fee, totalAmount, approvalStatus, withdrawalStatus, accountNumber, accountBank, accountHolder, createdAt]
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        amount:
          type: number
        fee:
          type: number
        totalAmount:
          type: number
        approvalStatus:
          type: string
          description: "pending | approved | rejected"
        withdrawalStatus:
          type: string
          description: "unpaid | paid"
        accountNumber:
          type: string
        accountBank:
          type: string
        accountHolder:
          type: string
        rejectReason:
          type: ["string", "null"]
        memberName:
          type: ["string", "null"]
        storeName:
          type: ["string", "null"]
        createdAt:
          type: string
    WithdrawalList:
      type: object
      required: [items, total, totalAmount, totalFee]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        totalAmount:
          type: number
        totalFee:
          type: number
    WithdrawalInput:
      type: object
      required: [amount, accountNumber, accountBank, accountHolder]
      properties:
        amount:
          type: number
        accountNumber:
          type: string
        accountBank:
          type: string
        accountHolder:
          type: string
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...