# lib/api-zod/src/generated/types/withdrawalList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TodoPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Withdrawal } from "./withdrawal";

export interface WithdrawalList {
  items: Withdrawal[];
  total: number;
  totalAmount: number;
  totalFee: number;
}